`EventManager` improvements and unit tests (#72)
I've made two improvements to `EventManager`:
- it now waits for the platform events websocket connection during
initialization (before, if it could not connect to the websocket, it
would initialize without errors, and then raise an error immediately
after)
- it now accepts floats as the timeout when waiting for event listeners
to complete

I've also written unit tests for it, since we didn't have it tested yet
and this was the third time I found some bug in it, which the tests
would catch if we had them.

+++ src/apify/actor.py
@@ -293,7 +293,7 @@ class Actor(metaclass=_ActorContextManager):
         cls,
         *,
         exit_code: int = 0,
-        event_listeners_timeout_secs: Optional[int] = EVENT_LISTENERS_TIMEOUT_SECS,
+        event_listeners_timeout_secs: Optional[float] = EVENT_LISTENERS_TIMEOUT_SECS,
     ) -> None:
         """Exit the actor instance.
 
@@ -305,7 +305,7 @@ class Actor(metaclass=_ActorContextManager):
 
         Args:
             exit_code (int, optional): The exit code with which the actor should fail (defaults to `0`).
-            event_listeners_timeout_secs (int, optional): How long should the actor wait for actor event listeners to finish before exiting
+            event_listeners_timeout_secs (float, optional): How long should the actor wait for actor event listeners to finish before exiting
         """
         return await cls._get_default_instance().exit(
             exit_code=exit_code,
@@ -316,7 +316,7 @@ class Actor(metaclass=_ActorContextManager):
         self,
         *,
         exit_code: int = 0,
-        event_listeners_timeout_secs: Optional[int] = EVENT_LISTENERS_TIMEOUT_SECS,
+        event_listeners_timeout_secs: Optional[float] = EVENT_LISTENERS_TIMEOUT_SECS,
     ) -> None:
         self._raise_if_not_initialized()
